[FIX] :: making try catch makes senses now

+++ src/main/java/br/com/exemplo/backendapitemplate/v1/service/UserService.java
@@ -18,9 +18,22 @@ public class UserService {
     }
 
     public User findById(Long id) {
-        return userRepository.findById(id).orElseThrow(
-                () -> new EntityNotFoundException("User not found")
-        );
+        // validating parameters received
+        if(id == null){
+            throw new IllegalArgumentException("Parâmetros inválidos ou nulos");
+        }
+
+        // if object exists
+        if(!userRepository.existsById(id)){
+            throw new EntityNotFoundException("Usuário não encontrado");
+        }
+
+        // mapping any errors
+        try{
+            return userRepository.findById(id).get();
+        }catch(EntityNotFoundException e){
+            throw new RuntimeException("Ocorreu um erro ao encontrar o usuário: "+e);
+        }
     }
 
     public User register(User user) {
@@ -28,11 +41,36 @@ public class UserService {
     }
 
     public User update(User user) {
-        return userRepository.save(user);
+        if(user.getId() == null){
+            throw new IllegalArgumentException("Parâmetros inválidos ou nulos");
+        }
+
+
+        if(!userRepository.existsById(user.getId())){
+            throw new EntityNotFoundException("Usuário não encontrado.");
+        }
+
+        try{
+            return userRepository.save(user);
+        }catch (Exception e){
+            throw new RuntimeException("Ocorreu um erro ao atualizar o usuário: "+e);
+        }
     }
 
     public void delete(Long id) {
-        userRepository.deleteById(id);
+        if(id == null){
+            throw new EntityNotFoundException("Parâmetros inválidos ou nulos");
+        }
+
+        if(!userRepository.existsById(id)){
+            throw new EntityNotFoundException("Usuário não encontrado.");
+        }
+
+        try{
+            userRepository.deleteById(id);
+        }catch (Exception e){
+            throw new RuntimeException("Ocorreu um erro ao deletar o usuário: "+e);
+        }
     }
